+++ jcpp-app/src/main/java/sanbing/jcpp/app/dal/entity/Order.java
@@ -59,7 +59,7 @@ public class Order implements Serializable {
 
     private String plateNo;
 
-    private Long settlementAmount;
+    private BigDecimal settlementAmount;
 
     private JsonNode settlementDetails;
